[MI-2999]:Fixed review comments
