+++ server/remote/msgraph/get_schedule_batched.go
@@ -44,7 +44,7 @@ type getScheduleRequestParams struct {
 
 func (c *client) GetSchedule(requests []*remote.ScheduleUserInfo, startTime, endTime *serializer.DateTime, availabilityViewInterval int) ([]*remote.ScheduleInformation, error) {
 	if !c.CheckUserStatus() {
-		c.Logger.Warnf(LogUserInActive, c.mattermostUserID)
+		c.Logger.Warnf(LogUserInactive, c.mattermostUserID)
 		return nil, errors.New(ErrorUserInActive)
 	}
