+++ server/serializer/event.go
@@ -1,12 +1,5 @@
 package serializer
 
-type User struct {
-	ID                string `json:"id"`
-	DisplayName       string `json:"displayName,omitempty"`
-	UserPrincipalName string `json:"userPrincipalName,omitempty"`
-	Mail              string `json:"mail,omitempty"`
-}
-
 type Event struct {
 	Start                      *DateTime            `json:"start,omitempty"`
 	Location                   *Location            `json:"location,omitempty"`
