+++ server/serializer/user.go
@@ -1,0 +1,8 @@
+package serializer
+
+type User struct {
+	ID                string `json:"id"`
+	DisplayName       string `json:"displayName,omitempty"`
+	UserPrincipalName string `json:"userPrincipalName,omitempty"`
+	Mail              string `json:"mail,omitempty"`
+}
